Blacklist std::experimental::optional when compiling with C++11
Both libstdc++ and libc++’s implementations of
std::experimental::optional require C++14, and will fail to compile in
C++11 mode. However, the headers are still present so the __has_include
check still passes.

This commit specifically blacklists using the standard versions if the
compiler is GCC or Clang and C++14 is not available.