Don't use libstdc++'s std::experimental::any in C++11 mode
The version of std::experimental::any in libstdc++ requires C++14, and
will static_assert if the compiler is in C++11 mode. So detect that and
use our override in that case.

(Note that this commit actually checks for a GCC-compatible compiler,
not libstdc++ specifically, so it would actually block C++11
implementation of std::any in libc++. But there isn’t one at the
moment, so we’ll opt to keep the check as simple as possible.)

This is a partial fix for issue #7