Don't use libstdc++'s std::experimental::string_view in C++11
As with any, libstdc++’s std::experimental::string_view requires C++14
and will static_assert in C++11. However, in this case libc++’s
string_view works fine in C++11 mode, so we need a slightly more
complicated check.

Specifically, if compiling in C++11 mode and the standard override
hasn’t already been applied, we’ll look for the __GLIBCXX__ symbol
defined by libstdc++. This requires opening a standard library header;
I chose <cstddef> because it’s relatively “small” and required by the
string_view implementation anyway.

This is the second half of a fix for issue #7