Fix override check for std::experimental::optional on Apple systems
libc++ on Apple has a missing symbol in the dynamic library
(specifically, the destructor for
std::experimental::bad_optional_access) which causes failures at link
time. This appears to be the case with the system libc++.dylib, and
upstream LLVM 3.7 and 3.8 (at least the versions available via
Homebrew).

This commit checks for the libc++ version by opening the <ciso646>
header, which is otherwise empty on C++. The check disables the
standard version if the libc++ version is <= 3.8; once 3.9 is released,
we’ll have to revisit this and see if it needs to be blacklisted too.

This is an updated fix for issue #5