Vastly increase test case coverage
Firstly, this commit enables CMake’s built-in testing environment. All
tests can now be run by calling “ctest” in the build directory.

Most tests are now built multiple times, in both C++11 and C++14 mode,
with forced overrides and without. The exceptions are the
implementation-specific tests for optional, any and variant, which do
not work with the standard versions. In addition, the variant test is
only built in C++14 mode.

Two extra tests, any_basic and optional_basic are included, which (as
the name suggest) are basic tests to ensure that we can compile and
link both with overrides and without.

This addresses issue #6

The all_includes test is currently commented out as this is broken;
that’s addressed in a separate bug.